Align deletion method signatures

+++ common/src/main/kotlin/dev/capybaralabs/d4j/store/common/repository/ChannelRepository.kt
@@ -8,8 +8,8 @@ interface ChannelRepository {
 	fun save(channel: ChannelData, shardId: Int): Mono<Void>
 	fun saveAll(channels: List<ChannelData>, shardId: Int): Mono<Void>
 
-	fun delete(channelId: Long, guildId: Long?): Mono<Int>
-	fun deleteByGuildId(channelIds: List<Long>, guildId: Long): Mono<Int>
+	fun delete(channelId: Long, guildId: Long?): Mono<Long>
+	fun deleteByGuildId(channelIds: List<Long>, guildId: Long): Mono<Long>
 	fun deleteByShardId(shardId: Int): Mono<Long>
 
 	fun countChannels(): Mono<Long>
