+++ common/src/main/kotlin/dev/capybaralabs/d4j/store/common/repository/MemberRepository.kt
@@ -10,9 +10,9 @@ interface MemberRepository {
 	// TODO we are potentially duplicating .user() data here, is there a way to avoid it?
 	fun saveAll(guildId: Long, members: List<MemberData>, shardId: Int): Mono<Void>
 
-	fun deleteById(guildId: Long, userId: Long): Mono<Int>
-	fun deleteByGuildId(guildId: Long): Mono<Int>
-	fun deleteByShardId(shardId: Int): Mono<Int>
+	fun deleteById(guildId: Long, userId: Long): Mono<Long>
+	fun deleteByGuildId(guildId: Long): Mono<Long>
+	fun deleteByShardId(shardId: Int): Mono<Long>
 
 	fun countMembers(): Mono<Long>
 	fun countMembersInGuild(guildId: Long): Mono<Long>
