+++ common/src/main/kotlin/dev/capybaralabs/d4j/store/common/repository/MessageRepository.kt
@@ -7,11 +7,11 @@ import reactor.core.publisher.Mono
 interface MessageRepository {
 	fun save(message: MessageData, shardId: Int): Mono<Void>
 
-	fun delete(messageId: Long): Mono<Int>
-	fun deleteByIds(messageIds: List<Long>): Mono<Int>
-	fun deleteByShardId(shardId: Int): Mono<Int>
-	fun deleteByChannelId(channelId: Long): Mono<Int>
-	fun deleteByChannelIds(channelIds: List<Long>): Mono<Int>
+	fun delete(messageId: Long): Mono<Long>
+	fun deleteByIds(messageIds: List<Long>): Mono<Long>
+	fun deleteByShardId(shardId: Int): Mono<Long>
+	fun deleteByChannelId(channelId: Long): Mono<Long>
+	fun deleteByChannelIds(channelIds: List<Long>): Mono<Long>
 
 	fun countMessages(): Mono<Long>
 	fun countMessagesInChannel(channelId: Long): Mono<Long>
