+++ common/src/main/kotlin/dev/capybaralabs/d4j/store/common/repository/UserRepository.kt
@@ -8,7 +8,7 @@ interface UserRepository {
 	fun save(user: UserData): Mono<Void>
 	fun saveAll(users: List<UserData>): Mono<Void>
 
-	fun deleteById(userId: Long): Mono<Int>
+	fun deleteById(userId: Long): Mono<Long>
 
 	fun countUsers(): Mono<Long>
