+++ postgres/src/main/kotlin/dev/capybaralabs/d4j/store/postgres/repository/PostgresMessageRepository.kt
@@ -1,6 +1,7 @@
 package dev.capybaralabs.d4j.store.postgres.repository
 
 import dev.capybaralabs.d4j.store.common.repository.MessageRepository
+import dev.capybaralabs.d4j.store.common.toLong
 import dev.capybaralabs.d4j.store.postgres.PostgresSerde
 import dev.capybaralabs.d4j.store.postgres.deserializeManyFromData
 import dev.capybaralabs.d4j.store.postgres.deserializeOneFromData
@@ -54,52 +55,52 @@ internal class PostgresMessageRepository(private val factory: ConnectionFactory,
 		}
 	}
 
-	override fun delete(messageId: Long): Mono<Int> {
+	override fun delete(messageId: Long): Mono<Long> {
 		return Mono.defer {
 			withConnection(factory) {
 				it.createStatement("DELETE FROM d4j_discord_message WHERE message_id = $1")
 					.bind("$1", messageId)
-					.executeConsumingSingle()
+					.executeConsumingSingle().toLong()
 			}
 		}
 	}
 
-	override fun deleteByIds(messageIds: List<Long>): Mono<Int> {
+	override fun deleteByIds(messageIds: List<Long>): Mono<Long> {
 		return Mono.defer {
 			withConnection(factory) {
 				it.createStatement("DELETE FROM d4j_discord_message WHERE message_id = ANY($1)")
 					.bind("$1", messageIds.toTypedArray())
-					.executeConsumingSingle()
+					.executeConsumingSingle().toLong()
 			}
 		}
 	}
 
-	override fun deleteByShardId(shardId: Int): Mono<Int> {
+	override fun deleteByShardId(shardId: Int): Mono<Long> {
 		return Mono.defer {
 			withConnection(factory) {
 				it.createStatement("DELETE FROM d4j_discord_message WHERE shard_index = $1")
 					.bind("$1", shardId)
-					.executeConsumingSingle()
+					.executeConsumingSingle().toLong()
 			}
 		}
 	}
 
-	override fun deleteByChannelId(channelId: Long): Mono<Int> {
+	override fun deleteByChannelId(channelId: Long): Mono<Long> {
 		return Mono.defer {
 			withConnection(factory) {
 				it.createStatement("DELETE FROM d4j_discord_message WHERE channel_id = $1")
 					.bind("$1", channelId)
-					.executeConsumingSingle()
+					.executeConsumingSingle().toLong()
 			}
 		}
 	}
 
-	override fun deleteByChannelIds(channelIds: List<Long>): Mono<Int> {
+	override fun deleteByChannelIds(channelIds: List<Long>): Mono<Long> {
 		return Mono.defer {
 			withConnection(factory) {
 				it.createStatement("DELETE FROM d4j_discord_message WHERE channel_id = ANY($1)")
 					.bind("$1", channelIds.toTypedArray())
-					.executeConsumingSingle()
+					.executeConsumingSingle().toLong()
 			}
 		}
 	}
